+++ src/node-view/types.ts
@@ -1,6 +1,0 @@
-export interface Node {
-  _internalId: string;
-  hidden: boolean;
-  opened: boolean;
-  [key: string]: any;
-}